feature: refactor rec-sys-app for python package
