+++ backend/Containerfile
@@ -1,6 +1,4 @@
-FROM quay.io/ecosystem-appeng/rec-sys-app:v1.0.1 AS algo-repo
-
-FROM registry.access.redhat.com/ubi9/python-311
+FROM registry.access.redhat.com/ubi9/python-312
 
 USER root
 WORKDIR /app/
@@ -9,17 +7,10 @@ RUN dnf update -y
 
 RUN pip3 install uv
 COPY pyproject.toml pyproject.toml
-RUN uv pip install -r pyproject.toml
+RUN uv sync
 
 # Copy the rest of the app
 COPY . .
-
-#TODO make sure feast fetching is possible
-# Copy the models and feature store from the algo-repo
-COPY --from=algo-repo /app/models/ public/models/
-COPY --from=algo-repo /app/service/ public/service/
-COPY --from=algo-repo /app/feature_repo/feature_store.yaml public/feature_store.yaml
-COPY --from=algo-repo /app/feature_repo/data public/data
 
 RUN chmod -R +r . && ls -la
